+++ app/server/modules/account-manager.js
@@ -1,38 +1,18 @@
 
-var crypto 		= require('crypto');
-var MongoDB 	= require('mongodb').Db;
-var Server 		= require('mongodb').Server;
-var moment 		= require('moment');
+const crypto 		= require('crypto');
+const moment 		= require('moment');
+const MongoClient 	= require('mongodb').MongoClient;
 
-/*
-	ESTABLISH DATABASE CONNECTION
-*/
-
-var dbName = process.env.DB_NAME || 'node-login';
-var dbHost = process.env.DB_HOST || 'localhost'
-var dbPort = process.env.DB_PORT || 27017;
-
-var db = new MongoDB(dbName, new Server(dbHost, dbPort, {auto_reconnect: true}), {w: 1});
-db.open(function(e, d){
-	if (e) {
+var db, accounts;
+MongoClient.connect(process.env.DB_URL, function(e, client) {
+	if (e){
 		console.log(e);
-	} else {
-		if (process.env.NODE_ENV == 'live') {
-			db.authenticate(process.env.DB_USER, process.env.DB_PASS, function(e, res) {
-				if (e) {
-					console.log('mongo :: error: not authenticated', e);
-				}
-				else {
-					console.log('mongo :: authenticated and connected to database :: "'+dbName+'"');
-				}
-			});
-		}	else{
-			console.log('mongo :: connected to database :: "'+dbName+'"');
-		}
+	}	else{
+		db = client.db(process.env.DB_NAME);
+		accounts = db.collection('accounts');
+		console.log('mongo :: connected to database :: "'+process.env.DB_NAME+'"');
 	}
 });
-
-var accounts = db.collection('accounts');
 
 /* login validation methods */